+++ src/main/java/com/amashchenko/maven/plugin/gitflow/GitFlowHotfixFinishMojo.java
@@ -52,6 +52,14 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
     @Parameter(property = "skipTestProject", defaultValue = "false")
     private boolean skipTestProject = false;
 
+    /**
+     * Whether to push to the remote.
+     * 
+     * @since 1.3.0
+     */
+    @Parameter(property = "pushRemote", defaultValue = "true")
+    private boolean pushRemote;
+
     /** {@inheritDoc} */
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
@@ -114,6 +122,8 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
 
             // fetch and check remote
             if (fetchRemote) {
+                gitFetchRemoteAndCompare(hotfixBranchName);
+
                 if (supportBranchName != null) {
                     gitFetchRemoteAndCompare(supportBranchName);
                 } else {
@@ -220,6 +230,10 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
                             && notSameProdDevName()) {
                         gitPush(gitFlowConfig.getDevelopmentBranch(), !skipTag);
                     }
+                }
+
+                if (!keepBranch) {
+                    gitPushDelete(hotfixBranchName);
                 }
             }
         } catch (CommandLineException e) {