+++ src/main/java/com/amashchenko/maven/plugin/gitflow/GitFlowSupportStartMojo.java
@@ -20,6 +20,7 @@ import java.util.Arrays;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.plugins.annotations.Parameter;
 import org.codehaus.plexus.components.interactivity.PrompterException;
 import org.codehaus.plexus.util.StringUtils;
 import org.codehaus.plexus.util.cli.CommandLineException;
@@ -27,9 +28,18 @@ import org.codehaus.plexus.util.cli.CommandLineException;
 /**
  * The git flow support start mojo.
  * 
+ * @since 1.5.0
  */
 @Mojo(name = "support-start", aggregator = true)
 public class GitFlowSupportStartMojo extends AbstractGitFlowMojo {
+
+    /**
+     * Whether to push to the remote.
+     * 
+     * @since 1.6.0
+     */
+    @Parameter(property = "pushRemote", defaultValue = "true")
+    private boolean pushRemote;
 
     /** {@inheritDoc} */
     @Override
@@ -72,6 +82,10 @@ public class GitFlowSupportStartMojo extends AbstractGitFlowMojo {
             if (installProject) {
                 // mvn clean install
                 mvnCleanInstall();
+            }
+
+            if (pushRemote) {
+                gitPush(gitFlowConfig.getSupportBranchPrefix() + tagName, false);
             }
         } catch (CommandLineException e) {
             getLog().error(e);